chatmind/pipeline/requirements.txt
# ChatMind Pipeline Requirements
# Core dependencies for the modular pipeline processing

# Data processing
pandas>=2.2.0
numpy>=1.24.0
jsonlines>=4.0.0

# Machine learning and embeddings
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
hdbscan>=0.8.33
umap-learn>=0.5.3

# AI/OpenAI (for cloud API methods)
openai>=1.0.0

# HTTP requests (for local model API calls)
requests>=2.31.0

# Database
neo4j>=5.0.0
qdrant-client>=1.7.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
click>=8.1.0

# Note: The pipeline also uses:
# - subprocess (built-in) for calling Ollama
# - pathlib (built-in) for file operations
# - json, logging, hashlib, pickle (built-in) for data handling
# - zipfile, re, unicodedata (built-in) for data processing 